requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.5
redis==6.4.0
rq==2.6.0
twilio==9.8.0
openai==1.107.2
pydantic-settings==2.10.1
python-multipart==0.0.20
PyYAML==6.0.2
pytest==8.4.1
httpx==0.28.1
